Revert "ignore stat calls, rather than treating them as reads"
This reverts commit 8d21f91a05392e37bed9de7190dd69cb0f37a4d3.

The reason is that I found that ignoring stat calls made fac rebuilds
far too fragile, particularly when using python, which caches a lot of
data in .pyc files.  Ignoring stats all too often meant we failed to
notice when we needed to rerun commands.  If python was only ever run
from fac, one might work around this, but it was too hard to maintain
that dependency information.