test linkat (and fix bug in linkat handling) and paths exceeding PATH_MAX
Make functions robust with respect to large path handling and NULL
pointers as in put.  This is not a thorough test of long paths, but it
did reveal a lot of problems.

Note that we are now ignoring long paths.  It might be better to
return an error somehow, but bigbro actually isn't well set up to
return error results.  :(